+++ R/distribute.R
@@ -2,8 +2,12 @@
 #'
 #' This function spatially distributes a set of sf features.
 #'
-#' The spatial features should be centred around 0,0. Use `align` to align different
-#' sets of simple spatial features around 0,0.
+#' The spatial features should be aligned around 0,0 for meaningful results.
+#' First use `converge` to bring together diversely distributed polygons around a single origin.
+#' Use `align` to align different sets of simple spatial features around the origin.
+#'
+#' Note that to rotate or scale features individually, you can use the `rotate` and `scale`
+#' methods with a vector parameters.
 #'
 #' @param x An sf-compatible feature layer, often containing polygons whose size is to be visually compared; REQUIRED.
 #' @param preserve.parameters Whether to preserve non-geometry parameters; default T.
@@ -15,8 +19,8 @@
 #' @param margin Scalar coefficient of spacing between features; default 1.2.
 #' @param x.mar Scalar coefficient of spacing between x values; default 1.
 #' @param y.mar Scalar coefficient of spacing between x values; default 1.
-#' @param scale Affine scaling of feature dimensions; default 1.
-#' @param angle Affine rotation of feature in degrees; default 0.
+#' @param scale Affine linear scaling of ALL feature dimensions, this is not areal scaling; default 1.
+#' @param angle Affine rotation of ALL feature in degrees; default 0.
 #' @seealso \code{\link{converge}}
 #' @return An sf object containing one or more features (with no defined CRS)
 #' @example
@@ -41,7 +45,7 @@ distribute <- function(x, preserve.parameters=T,
     warning("Friendly advice: distributing too many features may make it difficult to perceive size differences")
   }
   if (max.features<nrow(x)) {
-    warning(paste0("There are more features than the value of max.features; only the first ",
+    warning(paste0("There are more features (",nrow(x),") than the value of max.features; only the first ",
                    max.features," features will be distributed"))
     x <- x[1:max.features,]
   }
@@ -52,6 +56,7 @@ distribute <- function(x, preserve.parameters=T,
   bbox <- as.data.frame(do.call(rbind,bbox))
   bbox$diff_x <- bbox$xmax - bbox$xmin
   bbox$diff_y <- bbox$ymax - bbox$ymin
+
   widest_bb   <- max(bbox$diff_x)
   tallest_bb  <- max(bbox$diff_y)
   thinnest_bb <- min(bbox$diff_x)
@@ -86,7 +91,7 @@ distribute <- function(x, preserve.parameters=T,
         pos_y <- i %% rows
       }
       # Rotation should be done before other translation to ensure it is done around origin
-      rotate = function(a){
+      rotate_fun = function(a){
         r = a * pi / 180 #degrees to radians
         matrix(c(cos(r), sin(r), -sin(r), cos(r)), nrow = 2, ncol = 2)
       }
@@ -94,10 +99,10 @@ distribute <- function(x, preserve.parameters=T,
       if(isTRUE(preserve.parameters)) {
         d[[i]] <- sf::st_sf(
           data.frame( sf::st_drop_geometry(x[i,]),
-                      geom=sf::st_sf( (sf::st_geometry(x[i,]) * scale * rotate(angle)) + affine_transform) )
+                      geom=sf::st_sf( (sf::st_geometry(x[i,]) * scale * rotate_fun(angle)) + affine_transform) )
           )
       } else {
-        d[[i]] <- sf::st_sf((sf::st_geometry(x[i,]) * scale * rotate(angle)) + affine_transform)
+        d[[i]] <- sf::st_sf((sf::st_geometry(x[i,]) * scale * rotate_fun(angle)) + affine_transform)
       }
     }
     d <- sf::st_sf(do.call(rbind,d))