+++ NEWS.md
@@ -1,8 +1,15 @@
+# morphogram 0.3.0-beta
+
+- Rename of package to morphogram
+- Implement `align()`
+- Initial implementation of `rotate()`
+- Basic logo created
+
 # spatialcompare 0.2.0-beta
 
-- Reorganisation of functions, `converge` is now the main function name rather than "align" or "normalizeAroundCentroid"
-- Some bugs in `distribute` addressed with not taking bbox into account
+- Reorganisation of functions, `converge()` is now the main function name rather than "align" or "normalizeAroundCentroid"
+- Some bugs in `distribute()` addressed with not taking bbox into account
 
 # spatialcompare 0.1.0-beta
 
-- First release; with single form of `distribute` method (`regulargrid`)
+- First release; with single form of `distribute()` method (`regulargrid`)